+++ src/components/OnboardingSinglePage.tsx
@@ -11,6 +11,11 @@ import { CTA, OnboardingPage, renderOnboardingPage } from './Onboarding';
 const styles = StyleSheet.create({
   Wrapper: {
     flex: 1,
+    position: 'absolute',
+    top: 0,
+    right: 0,
+    bottom: 0,
+    left: 0,
   } as ViewStyle,
   ScrollView: {
     paddingBottom: bottomPanelStyles.BottomPanel.height,