+++ src/components/PaymentPanel.tsx
@@ -102,7 +102,7 @@ type PaymentPanelProps = {
   cancel: string;
   buttonOnPress: () => any;
   onIphoneX: boolean;
-  countBottomTabs?: boolean;
+  accountForBottomTabs?: boolean;
 };
 
 const transition = (
@@ -128,7 +128,7 @@ export const PaymentPanelUnmemoized = (props: PaymentPanelProps) => {
       {props.show && (
         <View
           style={
-            props.onIphoneX && !props.countBottomTabs
+            props.onIphoneX && !props.accountForBottomTabs
               ? [styles.WrapperChild, { height: 380 + IPHONE_X_HOME_AREA_HEIGHT }]
               : styles.WrapperChild
           }
@@ -162,7 +162,7 @@ export const PaymentPanelUnmemoized = (props: PaymentPanelProps) => {
               </View>
               <View
                 style={
-                  props.onIphoneX && !props.countBottomTabs
+                  props.onIphoneX && !props.accountForBottomTabs
                     ? [
                         styles.Footer,
                         { marginBottom: footerMarginBottom + IPHONE_X_HOME_AREA_HEIGHT },